backend/requirements.txt
django>=4.2,<5.0
djangorestframework>=3.14
django-cors-headers>=4.0
python-dotenv>=1.0
openai>=1.0
duckduckgo-search>=5.0
beautifulsoup4>=4.12
requests>=2.31
pydantic-settings>=2.0
diagrams>=0.23
playwright>=1.40
